Adds mapping to update POCL payload. Moves create POCL validaton error into pocl related file
