+++ packages/sales-api-service/src/services/exceptions/exceptions.service.js
@@ -1,9 +1,6 @@
 import { persist, StagingException, PoclStagingException } from '@defra-fish/dynamics-lib'
-import db from 'debug'
 import { getGlobalOptionSetValue } from '../reference-data.service.js'
-
-// @IWTF-2174: Remove after testing
-import { PoclValidationError } from './temp/index.js'
+import db from 'debug'
 const debug = db('sales:exceptions')
 
 /**
@@ -64,43 +61,4 @@ export const createTransactionFileException = async transactionFileError => {
   stagingException.bindToAlternateKey(PoclStagingException.definition.relationships.poclFile, transactionFileError.transactionFile)
   await persist([stagingException])
   return stagingException
-}
-
-const getPaymentData = async payment => {
-  const { source, method, ...rest } = payment
-  return {
-    ...rest,
-    paymentSource: source,
-    methodOfPayment: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.methodOfPayment.ref, method)
-  }
-}
-
-/**
- * @typedef {Object} TransactionValidationError
- * @property {!object} createTransactionPayload the data used to create a transaction
- * @property {!object} finaliseTransactionPayload the transaction data
- *
- * @param {TransactionValidationError} record
- * @returns {Promise<PoclValidationError>}
- */
-export const createDataValidationError = async record => {
-  debug('Adding exception for POCL record: %o', record)
-  const { dataSource, serialNumber, permissions: [permission] } = record.createTransactionPayload
-  const { licensee, issueDate: transactionDate, concessions, ...otherPermissionData } = permission
-  const validationErrorRecord = Object.assign(new PoclValidationError(), {
-    serialNumber,
-    transactionDate,
-    ...licensee,
-    ...otherPermissionData,
-    ...concessions && { concessions: JSON.stringify(concessions) },
-    ...await getPaymentData(record.finaliseTransactionPayload.payment),
-    status: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.status.ref, 'Needs Review'),
-    dataSource: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.dataSource.ref, dataSource),
-    preferredMethodOfConfirmation: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.preferredMethodOfConfirmation.ref, licensee.preferredMethodOfConfirmation),
-    preferredMethodOfNewsletter: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.preferredMethodOfNewsletter.ref, licensee.preferredMethodOfNewsletter),
-    preferredMethodOfReminder: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.preferredMethodOfReminder.ref, licensee.preferredMethodOfReminder)
-  })
-
-  await persist([validationErrorRecord])
-  return validationErrorRecord
 }
