+++ packages/sales-api-service/src/services/exceptions/pocl-data-validation-errors.service.js
@@ -1,5 +1,8 @@
 import { executeQuery, findById, persist, findPoclValidationErrors, PoclValidationError } from '@defra-fish/dynamics-lib'
 import { getGlobalOptionSetValue } from '../reference-data.service.js'
+import db from 'debug'
+import { Boom } from '@hapi/boom'
+const debug = db('sales:pocl-validation-errors')
 
 const POCL_VALIDATION_ERROR_STATUS_OPTIONSET = 'defra_poclvalidationerrorstatus'
 
@@ -14,9 +17,68 @@ export const getPoclValidationErrors = async () => {
   return results.map(result => result.entity)
 }
 
+const getPaymentData = async payment => {
+  const { source, method, ...rest } = payment
+  return {
+    ...rest,
+    paymentSource: source,
+    methodOfPayment: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.methodOfPayment.ref, method)
+  }
+}
+
+const getStatus = async record => {
+  const label = record.status ? record.status : 'Needs Review'
+  return getGlobalOptionSetValue(PoclValidationError.definition.mappings.status.ref, label)
+}
+
+const mapToPoclValidationErrorEntity = async record => {
+  const { dataSource, serialNumber, permissions: [permission] } = record.createTransactionPayload
+  const { licensee, issueDate: transactionDate, concessions, ...otherPermissionData } = permission
+  return Object.assign(new PoclValidationError(), {
+    serialNumber,
+    transactionDate,
+    ...licensee,
+    ...otherPermissionData,
+    ...concessions && { concessions: JSON.stringify(concessions) },
+    ...await getPaymentData(record.finaliseTransactionPayload.payment),
+    status: await getStatus,
+    dataSource: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.dataSource.ref, dataSource),
+    preferredMethodOfConfirmation: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.preferredMethodOfConfirmation.ref, licensee.preferredMethodOfConfirmation),
+    preferredMethodOfNewsletter: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.preferredMethodOfNewsletter.ref, licensee.preferredMethodOfNewsletter),
+    preferredMethodOfReminder: await getGlobalOptionSetValue(PoclValidationError.definition.mappings.preferredMethodOfReminder.ref, licensee.preferredMethodOfReminder)
+  })
+}
+
+/**
+ * @typedef {Object} TransactionValidationError
+ * @property {!object} createTransactionPayload the data used to create a transaction
+ * @property {!object} finaliseTransactionPayload the transaction data
+ *
+ * @param {TransactionValidationError} record
+ * @returns {Promise<PoclValidationError>}
+ */
+export const createDataValidationError = async record => {
+  debug('Adding exception for POCL record: %o', record)
+  const validationError = await mapToPoclValidationErrorEntity(record)
+  await persist([validationError])
+  return validationError
+}
+
+/**
+ * @typedef {Object} TransactionValidationError
+ * @property {!object} createTransactionPayload the data used to create a transaction
+ * @property {!object} finaliseTransactionPayload the transaction data
+ *
+ * @param {String} id - guid id of POCL validation error to update
+ * @param {TransactionValidationError} record
+ * @returns {Promise<PoclValidationError>}
+ */
 export const updatePoclValidationError = async (id, payload) => {
   const validationError = await findById(PoclValidationError, id)
-
-  const updated = Object.assign(validationError, payload)
+  if (!validationError) {
+    throw Boom.notFound('A POCL validation error with the given identifier could not be found')
+  }
+  const mappedRecord = await mapToPoclValidationErrorEntity(payload)
+  const updated = Object.assign(validationError, mappedRecord)
   await persist([updated])
 }
